+++ src/anemoi/inference/checkpoint/metadata/version_0_1_0.py
@@ -100,3 +100,13 @@ class Version_0_1_0(Metadata):
             dataset["attrs"] = dataset.copy()
 
         return ZarrRequest(dataset).graph(graph)
+
+    @property
+    def number_of_grid_points(self):
+        from .version_0_2_0 import ZarrRequest
+
+        dataset = self._dataset.copy()
+        if "attrs" not in dataset:
+            dataset["attrs"] = dataset.copy()
+
+        return ZarrRequest(dataset).number_of_grid_points