+++ CHANGELOG.md
@@ -12,11 +12,14 @@ Keep it human-readable, your future self will thank you!
 
 ### Added
 - ci: changelog release updater
+- earthkit-data replaces climetlab
 
 ### Changed
 - ci: updated workflows on PR and releases to use reusable actions
 
 ### Removed
+- climetlab
+
 
 ## [0.1.10] Fix missing constants